\name{eDMR.sub}
\alias{eDMR.sub}
\title{edmr sub function}
\usage{
eDMR.sub(myDiff, step = 100, dist = "none", DMC.qvalue = 0.01,
  DMC.methdiff = 25, num.DMCs = 1, num.CpGs = 3, DMR.methdiff = 20,
  granges = TRUE, plot = FALSE, main = "", direction = "both",
  ACF = TRUE, fuzzypval = 0.1)
}
\description{
edmr sub function
}

