\name{genebody.anno}
\alias{genebody.anno}
\title{generate genebody GRangesList object}
\usage{
genebody.anno(file)
}
\arguments{
  \item{bed}{file for gene models}
}
\description{
generate genebody GRangesList object
}
\examples{
genebody=genebody.anno(file="http://edmr.googlecode.com/files/hg19_refseq_all_types.bed")
}

