\name{getDMR}
\alias{getDMR}
\title{Calculate the p-value for the differentially methylated regions (ACF=TRUE)}
\usage{
getDMR(peaks, allMyDiff, pcutoff = 0.1, step = 100, DMC.qvalue = 0.01,
  DMC.methdiff = 25, num.DMCs = 1, num.CpGs = 3, DMR.methdiff = 20,
  a = TRUE)
}
\description{
Calculate the p-value for the differentially methylated
regions (ACF=TRUE)
}

