\name{myDiffToDMR}
\alias{myDiffToDMR}
\title{calculate differentially methylated regions from \code{myDiff} object.}
\usage{
myDiffToDMR(myDiff, dist = 100, step = 100, DMC.qvalue = 0.01,
  DMC.methdiff = 25, num.DMCs = 1, num.CpGs = 3, DMR.methdiff = 20,
  ACF = TRUE, fuzzypval = 0.1)
}
\description{
calculate differentially methylated regions from
\code{myDiff} object.
}

