\name{plotCost}
\alias{plotCost}
\title{plot cost function and the determined distance cutoff}
\usage{
plotCost(mixmdl, ...)
}
\arguments{
  \item{mixmdl}{object generated by
  \code{myDiff.to.mixmdl}}

  \item{main}{title for the plot}
}
\description{
plot cost function and the determined distance cutoff
}
\examples{
library(edmr)
library(mixtools)
data(edmr)

myMixmdl=myDiff.to.mixmdl(chr22.myDiff)
plotCost(myMixmdl, main="cost function")
}

