\name{plotdmrdistr}
\alias{plotdmrdistr}
\title{plot eDMR distribution over the subject genomic ranges}
\usage{
plotdmrdistr(myDMR, subject)
}
\arguments{
  \item{myDMR}{DMRs predicted by \code{edmr}.}

  \item{subject}{GRanges list used to annotate DMRs.}
}
\description{
plot eDMR distribution over the subject genomic ranges. This function has been deprecated.
}

