\name{plotMdl1}
\alias{plotMdl1}
\title{plot the fitted bimodal normal distribution for CpGs distances distribution}
\usage{
plotMdl1(mixmdl, subtitle = "", cex.sub = 1, ...)
}
\arguments{
  \item{mixmdl}{object generated by
  \code{myDiff.to.mixmdl}}

  \item{subtitle}{subtitle for the plot}

  \item{cex.sub}{subtitle font size}
}
\description{
plot the fitted bimodal normal distribution for CpGs
distances distribution
}
\examples{
library(edmr)
library(mixtools)
data(edmr)

myMixmdl=myDiff.to.mixmdl(myDiff)
plotMdl1(myMixmdl, subtitle="example", cex.sub=1.2)
}

