\name{stouffer.liptak}
\alias{stouffer.liptak}
\title{Stoffer Liptak test}
\usage{
stouffer.liptak(pvals, end, acf = acf, step = 100)
}
\arguments{
  \item{pvals}{p-values.}

  \item{end}{myDiff \code{end} column.}

  \item{acf}{output object from ACF function.}

  \item{step}{base pairs in each step of auto-correlation
  calculation.}
}
\description{
Stoffer Liptak test
}

