\name{plot.dmr.distr}
\alias{plot.dmr.distr}
\title{plot eDMR distribution over the subject genomic ranges}
\usage{
\method{plot}{dmr.distr}(myDMR, subject, ...)
}
\arguments{
  \item{myDMR}{DMRs predicted by \code{edmr}.}

  \item{GRanges}{list used to annotate DMRs.}

  \item{...}{Argument to be passed to \code{barplot}
  function.}
}
\description{
plot eDMR distribution over the subject genomic ranges
}

