\name{plotCost}
\alias{plotCost}
\title{plot cost function and the determined distance cutoff}
\usage{
plotCost(mixmdl, main = "")
}
\arguments{
  \item{mixmdl}{object generated by
  \code{myDiff.to.mixmdl}}

  \item{main}{plot tile}
}
\description{
plot cost function and the determined distance cutoff
}
\seealso{
\code{\link{myDiff.to.mixmdl}}
}

